Fixed type problem

+++ airline-data/src/main/scala/com/patson/model/LinkConsumptionHistory.scala
@@ -1,3 +1,3 @@
 package com.patson.model
 
-case class LinkConsumptionHistory(link : Link, passengerCount : Int, homeCountryCode : String, passengerType : PassengerType.Value, preferenceType : FlightPreferenceType.Value, preferredLinkClass : LinkClass, linkClass : LinkClass, satisfaction : Float)
+case class LinkConsumptionHistory(link : Link, passengerCount : Int, homeCountryCode : String, passengerType : PassengerType.Value, preferenceType : FlightPreferenceType.Value, preferredLinkClass : LinkClass, linkClass : LinkClass, satisfaction : Double)
